refactor: Update CommunityScreen component

+++ screens/DeleteAccScreen.js
@@ -1,17 +1,56 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, Text, TouchableOpacity, StyleSheet, Alert } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome5';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-const DeleteAccScreen = ({ navigation }) => {
+const API_URI = "https://whatever.lat/api/v1/"; // Adjust this to your actual API URL
 
+// Function to delete the current host
+const deleteCurrentHost = async (token) => {
+    try {
+        // Decode the token to extract the host ID
+        const decodedToken = jwt_decode(token);
+        const hostId = decodedToken.id; // Assuming the host ID is stored as 'id' in the token
+
+        // Log the hostId
+        console.log('Host ID:', hostId);
+
+        const response = await fetch(`${API_URI}/hosts/${hostId}`, {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`,
+            },
+        });
+        if (!response.ok) {
+            throw new Error("Failed to delete current host");
+        }
+
+        const json = await response.json();
+        return json;
+    } catch (error) {
+        console.error("Error deleting current host:", error);
+        throw error;
+    }
+};
+
+const DeleteAccScreen = ({ navigation }) => {
   const handleDeleteAccount = async () => {
     try {
+      const token = await AsyncStorage.getItem("token");
+      if (!token) {
+        throw new Error("No token found");
+      }
+      // Call the delete API
+      const deleteResponse = await deleteCurrentHost(token);
+      console.log('Delete Response:', deleteResponse);
+
       // Perform account deletion actions, e.g., delete user data
       await AsyncStorage.clear(); // Clear all stored data (adjust as per your app's logic)
       navigation.navigate('SignUp'); // Navigate to your login screen after deletion
     } catch (error) {
       console.error('Error deleting account:', error);
+      Alert.alert("Error", "Failed to delete account. Please try again.");
     }
   };
